over

+++ src/main/java/com/yangliuxin/controller/wechat/WeChatController.java
@@ -343,7 +343,7 @@ public class WeChatController {
         log.info(">>>>>>LOTTERY_REQUEST_DATA:{}", request);
         log.info(">>>>>>LOTTERY_BEAN_DATA:{}",lotteryBean);
         Lottery lottery = lotteryRepository.getById(lotteryBean.getId());
-        if(lottery == null || !users.getId().equals(lotteryBean.getUserId()) || !lottery.getUserId().equals(lotteryBean.getUserId())  || !lottery.getInd().equals(lotteryBean.getInd())){
+        if(lottery == null || !users.getId().equals(lotteryBean.getUserId().longValue()) || !lottery.getUserId().equals(lotteryBean.getUserId())  || !lottery.getInd().equals(lotteryBean.getInd())){
             log.info(">>>>>>LOTTERY_DB_DATA:{}",lottery);
             resultVo.setCode(0);
             resultVo.setMsg("不存在的中奖信息");
